new devise does not generate this, so this is not longer needed in migration

+++ db/migrate/20120428135420_devise_2_0.rb
@@ -1,6 +1,8 @@
 class Devise20 < ActiveRecord::Migration
   def up
-    remove_column :users, :remember_token
+    if column_exists? :users, :remember_token
+      remove_column :users, :remember_token
+    end
   end
 
   def down
